# SHA3: xor_64(varargs) -> xor_64_2(a,b), xor_64_5(a,b,c,d,e)