You already have some of these a b c d e functions.
Thought I'd try my luck doing the same for xor_64(), too.

Benchmark using your "SHA3-512 Long TEXT Input - HEX Output" test in Firefox 53 (three runs)

before: 865.08ms 862.54ms 864.88ms
after: 634.05ms 635.46ms 632.65ms

If this makes sense then xor_64(var_args) [now unused] could be removed.