@Caligatio 2017-04-03
If you can remove the old xor_64(varargs) in this PR, I'll be happy to accept.