@frostschutz 2017-04-03
Okay, I removed it. Also I noticed the Google Closure Compiler seems to remove xor_64_5 on its own accord (puts it directly in the code as far as I can tell), whereas xor_64_2 is retained as a function. Not sure how it makes that decision / what difference it makes. Thanks.